+++ SourceryTests/Parsing/FileParserSpec.swift
@@ -91,6 +91,42 @@ class FileParserSpec: QuickSpec {
                             "}")
                         expect(result.first).to(equal(expectedType))
                     }
+
+                    it("extracts annotations when declaration has an attribute on the preceding line") {
+                        let annotations = ["Annotation": NSNumber(value: true)]
+
+                        let expectedClass = Class(name: "SomeClass", variables: [], attributes: ["MainActor": [Attribute(name: "MainActor")]], annotations: annotations)
+                        let result = parse("""
+                        //sourcery: Annotation
+                        @MainActor
+                        class SomeClass {
+                        }
+                        """)
+                        expect(result.first).to(equal(expectedClass))
+                    }
+
+                    it("extracts annotations when declaration has a directive on the preceding line") {
+                        let annotations = ["Annotation": NSNumber(value: true)]
+                        let result = parse("""
+                        //sourcery: Annotation
+                        #warning("some warning")
+                        class SomeClass {
+                        }
+                        """)
+                        expect(result.first?.annotations).to(equal(annotations))
+                    }
+
+                    it("extracts annotations when declaration has a directive and an attribute on the preceding line") {
+                        let annotations = ["Annotation": NSNumber(value: true)]
+                        let result = parse("""
+                        //sourcery: Annotation
+                        #warning("some warning")
+                        @MainActor
+                        class SomeClass {
+                        }
+                        """)
+                        expect(result.first?.annotations).to(equal(annotations))
+                    }
                 }
 
                 context("given struct") {
@@ -323,6 +359,41 @@ class FileParserSpec: QuickSpec {
 
                         expect(parse("/// doc\n// sourcery: thirdLine = 4543\n/// comment\n// firstLine\n///baz\nclass Foo: TestProtocol { }", parseDocumentation: true))
                                 .to(equal([expectedType]))
+                    }
+
+                    it("extracts documentation correctly if there is a directive on preceeding line") {
+                        let expectedType = Class(name: "Foo", accessLevel: .internal, isExtension: false, variables: [], inheritedTypes: ["TestProtocol"])
+                        expectedType.annotations["thirdLine"] = NSNumber(value: 4543)
+                        expectedType.documentation = ["doc", "comment", "baz"]
+
+                        expect(parse("""
+                            /// doc
+                            // sourcery: thirdLine = 4543
+                            /// comment
+                            // firstLine
+                            ///baz
+                            #warning("a warning")
+                            class Foo: TestProtocol { }
+                            """, parseDocumentation: true))
+                        .to(equal([expectedType]))
+                    }
+
+                    it("extracts documentation correctly if there is a directive and an attribute on preceeding line") {
+                        let expectedType = Class(name: "Foo", accessLevel: .internal, isExtension: false, variables: [], inheritedTypes: ["TestProtocol"], attributes: ["MainActor": [Attribute(name: "MainActor")]])
+                        expectedType.annotations["thirdLine"] = NSNumber(value: 4543)
+                        expectedType.documentation = ["doc", "comment", "baz"]
+
+                        expect(parse("""
+                            /// doc
+                            // sourcery: thirdLine = 4543
+                            /// comment
+                            // firstLine
+                            ///baz
+                            #warning("a warning")
+                            @MainActor
+                            class Foo: TestProtocol { }
+                            """, parseDocumentation: true))
+                        .to(equal([expectedType]))
                     }
                 }